Simplify mode select to remove view-only mode
